Create static.yml
Welcome to my first website hosting experience using GitHub! I'm excited to share this journey of creating and deploying a website from scratch. With GitHub Pages, I'm learning the ins and outs of web hosting, version control, and showcasing my projects to the world. Follow along as I dive into the world of web development and bring my ideas to life online. Thanks for joining me on this adventure!